Split upgrade-progress feed out of routes_operations.py
CI's pytest check failed after the earlier commit here: the fixes to
get_upgrade_progress() pushed routes_operations.py to 419 lines, over
this package's 400-line guidance (tests/test_maintenance_package.py::
test_package_modules_stay_within_the_size_guidance). This package was
already split by concern from a single 1,802-line module specifically to
stay under that limit, so the fix is the same pattern every other
routes_*.py file here follows: move get_upgrade_progress() and the
log-tailing/classification helpers it alone uses into their own module,
routes_upgrade_progress.py, registered the same way every other route
module in this package is (imported for side effect in __init__.py).

routes_operations.py: 419 -> 252 lines. No behavior change -- purely a
file-boundary move, same reasoning both files carry inline.
